requirements.txt
anyio==4.3.0
black==24.4.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
docutils==0.21.2
fonttools==4.51.0
gmsh==4.11.1
h11==0.14.0
idna==3.7
jedi==0.19.1
kiwisolver==1.4.5
llvmlite==0.42.0
marimo==0.4.7
Markdown==3.6
matplotlib==3.8.4
mpmath==1.3.0
mypy-extensions==1.0.0
numba==0.59.1
numpy==1.26.4
packaging==24.0
parso==0.8.4
pathspec==0.12.1
pillow==10.3.0
platformdirs==4.2.1
py-pde==0.38.0
Pygments==2.17.2
pymdown-extensions==10.8.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
PyYAML==6.0.1
scipy==1.13.0
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
sympy==1.12
tomlkit==0.12.4
tqdm==4.66.2
uvicorn==0.29.0
websockets==12.0
